Rewrite the indexer to support real folders
Adds indexing of "real" folders so we can tie an actual filesystem path
to a folder in order to help in cases there are a lot of locations that
map to a single location.

Various parts of the indexer have been split into their own files to
make it easier to find logic for a given sub-component.  Various
"public-folder-only" assumptions had to be redesigned.

Note that this was tested by dumping all files, folders, and categories
from "master" and diffing against this commit.  These three tables were
precisely the same across over four million indexed files, so I'm very
confident this rewrite worked.